% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_large.R
\name{plot.lwdid_large}
\alias{plot.lwdid_large}
\title{Event-study plot for a large-N lwdid object}
\usage{
\method{plot}{lwdid_large}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lwdid_large"} from \code{lwdid(..., method=)}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots WATT(r) against relative time r with simultaneous (sup-t) confidence
bands. Requires the \pkg{ggplot2} package.
}
