% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_large.R
\name{print.lwdid_large}
\alias{print.lwdid_large}
\title{Print a large-N lwdid object}
\usage{
\method{print}{lwdid_large}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"lwdid_large"} from \code{lwdid(..., method=)}.}

\item{digits}{Number of digits for the results table.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a large-N lwdid object
}
