% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_large.R
\name{summary.lwdid_large}
\alias{summary.lwdid_large}
\title{Summarise a large-N lwdid object}
\usage{
\method{summary}{lwdid_large}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"lwdid_large"}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarise a large-N lwdid object
}
